moved classes related to the Trade up system into 'tradeup' package
