+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/tradeup/TradeUpValidator.java
@@ -11,7 +11,7 @@
  *              does given list of cards have the the requirement number of cards
  */
 
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.tradeup;
 
 import com.internetEnemies.combatCritters.objects.Card;
 import com.internetEnemies.combatCritters.objects.ItemStack;
