+++ app/src/test/java/com/internetEnemies/combatCritters/IntegrationTests/TradeUpHandlerIntegrationTest.java
@@ -10,11 +10,11 @@
 package com.internetEnemies.combatCritters.IntegrationTests;
 
 
-import com.internetEnemies.combatCritters.Logic.ITradeUpHandler;
-import com.internetEnemies.combatCritters.Logic.ITradeUpValidator;
+import com.internetEnemies.combatCritters.Logic.tradeup.ITradeUpHandler;
+import com.internetEnemies.combatCritters.Logic.tradeup.ITradeUpValidator;
 import com.internetEnemies.combatCritters.Logic.transaction.ITransactionHandler;
-import com.internetEnemies.combatCritters.Logic.TradeUpHandler;
-import com.internetEnemies.combatCritters.Logic.TradeUpValidator;
+import com.internetEnemies.combatCritters.Logic.tradeup.TradeUpHandler;
+import com.internetEnemies.combatCritters.Logic.tradeup.TradeUpValidator;
 import com.internetEnemies.combatCritters.Logic.transaction.TransactionHandler;
 import com.internetEnemies.combatCritters.Logic.exceptions.InvalidTradeUpCardsException;
 import com.internetEnemies.combatCritters.TestUtils;
